Move classes from "Controls/Tools" into "ToolsUI"
