+++ pwiz_tools/Skyline/Model/DdaSearch/MsgfPlusSearchEngine.cs
@@ -30,9 +30,9 @@ using System.Linq;
 using System.Text;
 using pwiz.Common.Collections;
 using pwiz.Common.SystemUtil;
-using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Model.Results;
 using pwiz.Skyline.Properties;
+using pwiz.Skyline.ToolsUI;
 
 namespace pwiz.Skyline.Model.DdaSearch
 {
