+++ pwiz_tools/Skyline/TestFunctional/PythonInstallerTest.cs
@@ -23,7 +23,6 @@ using System.Collections.ObjectModel;
 using System.Windows.Forms;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using pwiz.Skyline.Alerts;
-using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Properties;
 using pwiz.Skyline.ToolsUI;
 using pwiz.Skyline.Util;
