+++ pwiz_tools/Skyline/TestFunctional/ToolServiceTest.cs
@@ -22,8 +22,8 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using pwiz.Skyline;
-using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Model;
+using pwiz.Skyline.ToolsUI;
 using pwiz.Skyline.Util;
 using pwiz.SkylineTestUtil;
 using SkylineTool;
