+++ pwiz_tools/Skyline/TestTutorial/QuasarTutorialTest.cs
@@ -26,7 +26,6 @@ using System.Text;
 using System.Windows.Forms;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using pwiz.Skyline.Controls.Databinding;
-using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Model;
 using pwiz.Skyline.Model.DocSettings;
 using pwiz.Skyline.SettingsUI;
