+++ pwiz_tools/Skyline/TestUtil/AbstractFunctionalTestEx.cs
@@ -34,7 +34,6 @@ using pwiz.ProteowizardWrapper;
 using pwiz.Skyline.Alerts;
 using pwiz.Skyline.Controls.Databinding;
 using pwiz.Skyline.Controls.GroupComparison;
-using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.FileUI;
 using pwiz.Skyline.Model;
 using pwiz.Skyline.Model.DocSettings;
