+++ pwiz_tools/Skyline/ToolsUI/ToolInstallUI.cs
@@ -27,7 +27,6 @@ using System.Windows.Forms;
 using Newtonsoft.Json;
 using pwiz.Skyline.Alerts;
 using pwiz.Skyline.Controls;
-using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Model.Databinding;
 using pwiz.Skyline.Model.DocSettings;
 using pwiz.Skyline.Properties;
