+++ pwiz_tools/Skyline/Util/UtilUI.cs
@@ -27,10 +27,10 @@ using System.Threading;
 using System.Windows.Forms;
 using pwiz.Common.SystemUtil;
 using pwiz.Skyline.Alerts;
-using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.EditUI;
 using pwiz.Skyline.Model;
 using pwiz.Skyline.Properties;
+using pwiz.Skyline.ToolsUI;
 using ZedGraph;
 
 namespace pwiz.Skyline.Util
